Updates to allow request headers to be more realistic.
It seemed like requests were hanging, so I modified requests to include
realistic User-Agent field, as well as a CSRF token and a Session ID.
The latter are currently hardwired into the code, and will probably need
to be updated periodically.

Also, you can save request and response now in debug mode.